% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peat_hamstr.R
\name{peat_hamstr}
\alias{peat_hamstr}
\title{Estimates an age-depth model and parameters from a modified Clymo Bog Growth Model}
\usage{
peat_hamstr(
  depth,
  obs_age,
  obs_err,
  min_age = 1950 - as.numeric(format(Sys.Date(), "\%Y")),
  K = NULL,
  top_depth = NULL,
  bottom_depth = NULL,
  acc_mean_prior = NULL,
  acc_shape = 1.5,
  mem_mean = 0.5,
  mem_strength = 10,
  model_bioturbation = FALSE,
  n_ind = NULL,
  L_prior_mean = 10,
  L_prior_shape = 2,
  L_prior_sigma = NULL,
  model_displacement = FALSE,
  D_prior_scale = 10,
  model_hiatus = FALSE,
  H_top = NULL,
  H_bottom = NULL,
  layer_mass,
  cumulative_mass0,
  depth_clymo_par_constant,
  depth2_upper,
  depth2_lower,
  p1_layer_mass_shape = 5,
  p2_layer_mass_shape = 0.01,
  p3_layer_mass_shape = 100,
  p1_clymo_par = 3,
  p2_clymo_par = 0.02,
  p3_clymo_par = 100,
  p1_clymo_alpha_1 = 2,
  p2_clymo_alpha_1 = 50,
  p3_clymo_alpha_1 = 1/1000,
  p1_clymo_alpha_2 = 3,
  p2_clymo_alpha_2 = 3/0.005,
  p3_clymo_alpha_2 = 1/10000,
  p1_ac_age = 3,
  p2_ac_age = 3/80,
  p3_ac_age = 40,
  clymo_par_memory_p1 = 1,
  clymo_par_memory_p2 = 1,
  p1_age0 = 5,
  p2_age0 = 5/1,
  p3_age0 = 1,
  sample_posterior = TRUE,
  hamstr_control = list(),
  stan_sampler_args = list()
)
}
\arguments{
\item{depth}{depths of observed ages (age control points)}

\item{obs_age}{observed age at each depth (age control points)}

\item{obs_err}{error associated with each observed age (1 standard error)}

\item{min_age}{the minimum age that the first modelled depth can be. Useful
if extrapolating above the shallowest age control point to e.g. the
surface. So set min_age to the year the core was collected. E.g. for a core
collected in 1990, with ages in years BP this would be -40 (present = 1950
by convention). The default value is the current year in calendar age BP
(e.g. -71 for 2021).}

\item{K}{controls the number and structure of the hierarchically modelled
down-core sections. It is specified as a vector, where each value indicates
the number of new child sections for each parent section, e.g., c(10, 10,
10) would specify 10 sections at the coarsest level, with 10 new sections
at next finer level, giving a total of 1000 sections at the finest
resolution.

By default, the total number of sections at the finest level is set so that
for the median distance between age control points there are 16 sections.

By default a "powers of 2" hierarchical structure is chosen and then
adjusted to get close to the desired number of sections, e.g. for a desired
100 sections the result is c(2, 2, 2, 2, 2, 3), giving 96 sections.}

\item{top_depth, bottom_depth}{the top and bottom depths of the desired
age-depth model. Must encompass the range of the data. Defaults to the
shallowest and deepest data points.}

\item{acc_mean_prior}{hyperparameter for the prior on the overall mean
accumulation rate for the record. Units are obs_age / depth. E.g. if depth
is in cm and age in years then the accumulation rate is in years/cm. The
overall mean accumulation rate is given a weak half-normal prior with mean
= 0, SD = 10 * acc_mean_prior. If left blank, acc_mean_prior is set to the
mean accumulation rate estimated by fitting a robust linear model using
\link[MASS]{rlm}.}

\item{acc_shape}{hyperparameter for the shape of the priors on accumulation
rates. Defaults to 1.5 - as for Bacon 2.2.}

\item{mem_mean}{hyperparameter; a parameter of the Beta prior distribution on
"memory", i.e. the autocorrelation parameter in the underlying AR1 model.
The prior on the correlation between layers is scaled according to the
thickness of the sediment sections in the highest resolution hierarchical
layer, \emph{delta_c}, which is determined by the total length age-models and
the parameter vector \emph{K}. mem_mean sets the mean value for \emph{R} (defaults to
0.5), while \emph{w} = R^(delta_c)}

\item{mem_strength}{hyperparameter: sets the strength of the memory prior,
defaults to 10 as in Bacon >= 2.5.1}

\item{model_bioturbation}{defaults to FALSE. If TRUE, additional uncertainty
in the observed ages due to sediment mixing (bioturbation) is modelled via
a latent variable process. The amount of additional uncertainty is a
function of the mixing depth L, the sedimentation rate, and the number of
particles (e.g. individual foraminifera) per measured date. See description
for details.}

\item{n_ind}{the number of individual particles (e.g. foraminifera) in each
sample that was dated by e.g. radiocarbon dating. This can be a single
value or a vector the same length as obs_age. Only used if
model_bioturbation = TRUE.}

\item{L_prior_mean}{mean of the gamma prior on mixing depth, defaults to 10.}

\item{L_prior_shape, L_prior_sigma}{shape and standard deviation of the gamma
prior on the mixing depth. Set only one of these, the other will be
calculated. Defaults to shape = 2. If either the shape or sigma parameter
is set to zero, the mixing depth is fixed at the value of L_prior_mean,
rather than being sampled.}

\item{model_displacement}{model additional error on observed ages that does
not scale with the number of individual particles in a sample, for example
due to incomplete mixing.}

\item{D_prior_scale}{scale of the half-normal prior on additional error on
observed ages. The mean and standard deviation of a half-normal are equal
to the scale. Units are those of the depth variable, e.g. cm.}

\item{model_hiatus}{optionally model a hiatus.}

\item{H_top, H_bottom}{limits to the location of a hiatus. By default these
are set to the top and bottom data points but can be set by the user}

\item{cumulative_mass0}{A numeric value with the cumulative mass of all peat
layers above the first considered peat layer (kg m\eqn{^{-1}}).}

\item{depth2_upper, depth2_lower}{A numeric vector with the same length as
\code{cumulative_mass} defining the depth of the upper and lower boundaries of the
corresponding peat layers.}

\item{p1_layer_mass_shape, p2_layer_mass_shape}{A numeric
value representing the shape and rate parameter of a Gamma distribution
modeling the shape parameter of the Gamma distribution assumed for modeled
peat cumulative masses.}

\item{p1_clymo_par, p2_clymo_par}{A numeric
value representing the shape and rate parameter of the Gamma distribution
modeling peat mass addition rates (g m\eqn{^{-2}} yr\eqn{^{-1}}).}

\item{sample_posterior}{if set to FALSE, hamstr skips sampling the model and
returns only the data, model structure and prior parameters so that data
and prior distributions can be plotted and checked prior to running a
model. Defaults to TRUE}

\item{hamstr_control}{additional arguments to hamstr useful for debugging or
development. See \code{\link[hamstr]{hamstr_control}} for details.}

\item{stan_sampler_args}{additional arguments to \link[rstan]{sampling} passed as a
named list. e.g. list(chains = 8, iter = 4000) to run 8 MCMC chains of 4000
iterations instead of the default 4 chains of 2000 iterations.
See \code{\link[hamstr]{get_stan_sampler_args}} for details.}

\item{cumulative_mass}{A numeric vector with the cumulative mass of the peat
layers (kg m\eqn{^{-1}}).}

\item{depth2}{A numeric vector with the same length as \code{cumulative_mass}
defining the depth of the corresponding peat layers.}

\item{p1_clymo_alpha, p2_clymo_alpha}{A numeric
value representing the shape and rate parameter of the Gamma distribution
modeling peat decomposition rate (in a simple exponential decomposition
model) (yr\eqn{^{-1}}).}
}
\description{
Estimates an age-depth model and parameters from a modified Clymo Bog Growth Model
}
