% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peathamstr-package.R
\docType{package}
\name{peathamstr-package}
\alias{peathamstr-package}
\alias{peathamstr}
\title{The 'peathamstr' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
