% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peat_hamstr.R
\name{plot.peathamstr_fit}
\alias{plot.peathamstr_fit}
\title{Plots objects of class \code{peathamstr_fit}}
\usage{
\method{plot}{peathamstr_fit}(
  x,
  type = c("default", "age_models", "acc_rates", "hier_acc_rates",
    "acc_mean_prior_post", "mem_prior_post", "L_prior_post", "D_prior_post", "PDF_14C",
    "cumulative_mass_profile", "cumulative_carbon_mass_profile", "mass_fluxes",
    "carbon_fluxes"),
  summarise,
  carbon_content = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{peathamstr_fit}.}

\item{summarise}{See \code{hamstr::plot.hamstr_fit()}.}

\item{...}{Additional paramters (currently ignored).}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots objects of class \code{peathamstr_fit}
}
