% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summary.peathamstr_fit}
\alias{summary.peathamstr_fit}
\title{Summarizes objects of class \code{peathamstr_fit}}
\usage{
\method{summary}{peathamstr_fit}(
  object,
  type = c("age_models", "acc_rates", "pars", "mass_fluxes", "carbon_fluxes",
    "cumulative_mass"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{peathamstr_fit}.}

\item{...}{further parameters passed to individual summary functions.}
}
\value{
A tibble with summarized posterior draws.
}
\description{
Summarizes objects of class \code{peathamstr_fit}
}
