% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peat_hamstr.R
\name{plot.peathamstr_fit}
\alias{plot.peathamstr_fit}
\title{Plots objects of class \code{peathamstr_fit}}
\usage{
\method{plot}{peathamstr_fit}(
  x,
  type = c("default", "age_models", "acc_rates", "hier_acc_rates", "acc_mean_prior_post",
    "mem_prior_post", "L_prior_post", "D_prior_post", "PDF_14C",
    "cumulative_mass_profile", "mass_fluxes"),
  summarise,
  ...
)
}
\arguments{
\item{x}{An object of class \code{peathamstr_fit}.}

\item{type}{A character value. Must be one of the plot \code{type}s available for
\code{hamstr_fit} objects (see \code{hamstr::plot.hamstr_fit()}) or the additional types
provided by the 'peathamstr' package:
\describe{
\item{\code{"cumulative_mass_profile"}}{Plots the modeled and measured
cumulative masses versus depth.}
\item{\code{"mass_fluxes"}}{Plots the modeled and measured
net mass uptake (NMU) and net mass release (NMR) versus average estimated
ages. Shaded areas are 50\\% and 95\\% posterior intervals.}
}}

\item{summarise}{See \code{hamstr::plot.hamstr_fit()}.}

\item{...}{Additional paramters (currently ignored).}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots objects of class \code{peathamstr_fit}
}
