% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{predict.peathamstr_fit}
\alias{predict.peathamstr_fit}
\title{Predict method for objects of class \code{peathamstr_fit}}
\usage{
\method{predict}{peathamstr_fit}(
  object,
  type = c("age_models", "acc_rates", "mass_fluxes", "carbon_fluxes",
    "cumulative_mass", "cumulative_carbon_mass", "apparent_mass_accumulation_rates",
    "apparent_carbon_accumulation_rates"),
  depth = c("modelled", "data"),
  carbon_content = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{peathamstr_fit}.}

\item{depth}{A character value defining depth values for which predictions are
computed. Must be one of:
\describe{
\item{\code{"modelled"}}{Predicts for depths the model was constructed on.}
\item{\code{"data"}}{Predicts for depths of the input data to the model.}
}
Alternatively can be a numeric vector to specify depths (this only works if
\code{type == "age_models"}).}

\item{carbon_content}{A numeric vector with the same length as there are
\code{depth2} values in the input data. This only needs to be specified when
\code{type == "carbon_fluxes"} and otherwise can be set to \code{NULL}. This will be
used to interpolate carbon contents for layers the model was constructed on.
The interpolated carbon contents are multiplied with the mass fluxes (as
predicted when setting \code{type} to \code{"mass_fluxes"}) to compute respective
carbon fluxes.}

\item{...}{additional arguments to \code{hamstr} predict methods}
}
\value{
A \code{tibble} of \code{peathamstr_fit} age-depth model realizations.
}
\description{

}
