% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_methods.R
\name{summary.peathamstr_fit}
\alias{summary.peathamstr_fit}
\title{Summarizes objects of class \code{peathamstr_fit}}
\usage{
\method{summary}{peathamstr_fit}(
  object,
  type = c("age_models", "acc_rates", "pars", "mass_fluxes", "carbon_fluxes",
    "cumulative_mass"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{peathamstr_fit}.}

\item{type}{A character value defining what posterior variables are predicted.
One of:
\describe{
\item{\code{"age_models"}}{Predicts posterior ages at specified depths.}
\item{\code{"acc_rates"}}{Predicts posterior age accumulation rates
(yr cm\eqn{^{-1}}) at the depths the model was constructed on}
\item{\code{"mass_fluxes"}}{Predicts posterior mass fluxes (kg yr\eqn{^{-1}}
m\eqn{^{-2}}) at the depths the model was constructed on.}
\item{\code{"carbon_fluxes"}}{Predicts posterior carbon mass fluxes
(kg yr\eqn{^{-1}} m\eqn{^{-2}}) at the depths the model was constructed
on. For this, argument \code{carbon_content} needs to be specified.}
}}

\item{...}{further parameters passed to individual summary functions.}
}
\value{
A tibble with summarized posterior draws.
}
\description{
Summarizes objects of class \code{peathamstr_fit}
}
